♻️:코드컴포넌트화
