+++ src/Components/modal/HeartIcon.tsx
@@ -1,0 +1,40 @@
+import styled from "styled-components";
+
+interface HeartIconProps {
+  handleAddLikedList: () => void;
+  isFilled: boolean;
+}
+
+const HeartIcon = ({ handleAddLikedList, isFilled }: HeartIconProps) => {
+  return (
+    <InfoBtn onClick={handleAddLikedList}>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill={isFilled ? "#fff" : "none"}
+        viewBox="0 0 24 24"
+        stroke-width="1.5"
+        stroke="#fff"
+        className="size-6"
+      >
+        <path
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          d="M21 8.25c0-2.485-2.099-4.5-4.688-4.5-1.935 0-3.597 1.126-4.312 2.733-.715-1.607-2.377-2.733-4.313-2.733C5.1 3.75 3 5.765 3 8.25c0 7.22 9 12 9 12s9-4.78 9-12Z"
+        />
+      </svg>
+    </InfoBtn>
+  );
+};
+
+export default HeartIcon;
+
+// style
+const InfoBtn = styled.button`
+  width: 32px;
+  margin-right: 1rem;
+  cursor: pointer;
+  opacity: 0.7;
+  &:hover {
+    opacity: 1;
+  }
+`;
