+++ src/Components/movie-page/Slider.tsx
@@ -1,12 +1,10 @@
-import { AnimatePresence, motion } from "framer-motion";
 import styled from "styled-components";
-import { makeImagePath } from "../../utilities";
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
 import { IGetMoviesResult } from "../../api";
 
 import DoubleRight from "../../icon/double-right.svg";
+import SliderRow from "./SliderRow";
 
 interface SliderProps {
   category: {
@@ -16,35 +14,8 @@ interface SliderProps {
 }
 
 const offset = 6;
-const rowVariants = {
-  hidden: {
-    x: window.outerWidth,
-  },
-  visible: {
-    x: 0,
-  },
-  exit: {
-    x: -window.outerWidth,
-  },
-};
-const MovieVariants = {
-  normal: {
-    scale: 1,
-  },
-  hover: {
-    scale: 1.4,
-    y: -40,
-    transition: { delay: 0.5, duration: 0.3, type: "tween" },
-  },
-};
-const infoVariants = {
-  hover: {
-    opacity: 1,
-  },
-};
 
 const Slider = ({ category }: SliderProps) => {
-  const navigate = useNavigate();
   const [index, setIndex] = useState(0);
   const [leaving, setLeaving] = useState(false);
 
@@ -62,45 +33,18 @@ const Slider = ({ category }: SliderProps) => {
     setIndex((prev) => (prev === maxIndex ? 0 : prev + 1));
   };
 
-  const onClickMovie = (movieId: number) => {
-    navigate(`/movies/${movieId}`);
-    document.body.style.overflow = "hidden";
-  };
-
   if (isLoading) return null;
 
   return (
     <SliderWrapper>
       <SliderBox>
         <SliderTitle>{category.title}</SliderTitle>
-        <AnimatePresence initial={false} onExitComplete={toggleLeaving}>
-          <Row
-            key={`movies-${index}`}
-            variants={rowVariants}
-            initial="hidden"
-            animate="visible"
-            exit="exit"
-            transition={{ type: "tween", duration: 1 }}
-          >
-            {data?.results
-              .slice(offset * index, offset * index + offset)
-              .map((movie) => (
-                <Movie
-                  key={movie.id}
-                  bgImg={makeImagePath(movie.backdrop_path)}
-                  variants={MovieVariants}
-                  initial="normal"
-                  whileHover="hover"
-                  transition={{ type: "tween" }}
-                  onClick={() => onClickMovie(movie.id)}
-                >
-                  <MovieInfo variants={infoVariants}>
-                    <p>{movie.title}</p>
-                  </MovieInfo>
-                </Movie>
-              ))}
-          </Row>
-        </AnimatePresence>
+        <SliderRow
+          index={index}
+          data={data}
+          offset={offset}
+          toggleLeaving={toggleLeaving}
+        />
         <SliderBtn onClick={() => increaseIndex(data?.results.length || 0)}>
           <img src={DoubleRight} style={{ width: "32px" }} />
         </SliderBtn>
@@ -125,41 +69,7 @@ const SliderTitle = styled.h4`
   color: rgba(255, 255, 255, 0.9);
   font-size: 20px;
 `;
-const Row = styled(motion.div)`
-  width: 100%;
-  padding: 0 1rem;
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  gap: 0.5rem;
-  position: absolute;
-`;
-const Movie = styled(motion.div)<{ bgImg: string }>`
-  height: 140px;
-  border-radius: 0.5rem;
-  background-image: url(${(props) => props.bgImg});
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  cursor: pointer;
-  overflow: hidden;
-  &:first-child {
-    transform-origin: center left;
-  }
-  &:last-child {
-    transform-origin: center right;
-  }
-`;
-const MovieInfo = styled(motion.div)`
-  width: 100%;
-  padding: 1rem;
-  background: linear-gradient(transparent, #000);
-  opacity: 0;
-  position: absolute;
-  bottom: 0;
-  p {
-    font-weight: 600;
-  }
-`;
+
 const SliderBtn = styled.button`
   width: 40px;
   height: 160px;
